Added support for recording audio to and from internal storage

+++ app/src/main/java/com/zelyder/audionotes/media/exoplayer/MediaSource.kt
@@ -60,7 +60,7 @@ class MediaSource
                 )
                 .putLong(
                     MediaMetadataCompat.METADATA_KEY_DURATION,
-                    audio.duration.toLong()
+                    audio.duration
                 )
                 .build()
         }
